requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
backports.zoneinfo==0.2.1
billiard==4.2.1
black==24.8.0
blinker==1.8.2
Brotli==1.1.0
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==1.7
cron-descriptor==1.4.5
Django==4.2.18
django-celery-beat==2.7.0
django-timezone-field==7.1
djangorestframework==3.14.0
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
gevent==24.2.1
geventhttpclient==2.0.12
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
kombu==5.4.2
locust==2.25.0
MarkupSafe==2.1.5
msgpack==1.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
psutil==6.1.1
pydub==0.25.1
pytest==8.3.4
pytest-django==4.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.2.1
redis==5.2.1
requests==2.32.3
roundrobin==0.0.4
six==1.15.0
SpeechRecognition==3.10.4
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.6
zipp==3.20.2
zope.event==5.0
zope.interface==7.2
